Leave off point, in midst of refactoring
