+++ index.html
@@ -49,6 +49,7 @@
 	    <script type="text/javascript" src="./js/factory/keyframefactory.js"></script>
 	    <script type="text/javascript" src="./js/factory/gitdifffactory.js"></script>
 	    <script type="text/javascript" src="./js/factory/nwguifactory.js"></script>
+      <script type="text/javascript" src="./js/factory/dirTreeFactory.js"></script>
 	    <script type="text/javascript" src="./js/filters/datetime.js"></script>
 	    <script type="text/javascript" src="./js/ui-routes/projectbrowser/projectbrowser.js"></script>
 	    <script type="text/javascript" src="./js/ui-routes/scrubber/scrubber.js"></script>
