+++ js/dragdrop.js
@@ -16,6 +16,7 @@ dropbox.ondrop = function (e) {
   e.preventDefault();
 
   for (var i = 0; i < e.dataTransfer.files.length; ++i) {
+    window.globalVar = e.dataTransfer.files[i].path;
     console.log(e.dataTransfer.files[i].path);
   }
   return false;
