+++ js/ui-routes/projectbrowser/projectbrowser.js
@@ -6,9 +6,9 @@ app.directive('projectbrowser', function($rootScope) {
 	return {
 		restrict: 'E',
 		templateUrl: 'js/ui-routes/projectbrowser/projectbrowser.html',
-		controller: function ($scope, CommLinkFactory) {
+		controller: function ($scope, CommLinkFactory, DirTreeFactory) {
 
-			$scope.fileTree = require(path.join(process.env.PWD, "js", "dirTree.js"));
+			$scope.fileTree = DirTreeFactory.getTree(process.env.PWD);
 
 			// This watches for the latest file selected in the browser file tree
 			$scope.$watch('abc.currentNode.path', function(newFile, oldFile) {
